# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PetVet AI API
  description: |
    API for PetVet AI - A WhatsApp-first veterinary virtual assistant platform.

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    - API endpoints: 100 requests per 15 minutes per user
    - WhatsApp webhook: 1000 requests per minute
  version: 1.0.0
  contact:
    email: api@petvet.ai
  license:
    name: Proprietary
servers:
  - url: https://api.petvet.ai/api/v1
    description: Production server
  - url: https://staging-api.petvet.ai/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
...
  /reminders:
    get:
      tags:
        - Reminders
      summary: List reminders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, sent, acknowledged, snoozed]
        - name: upcoming
          in: query
          schema:
            type: boolean
          description: Filter to reminders in the next 7 days
      responses:
        '200':
          description: List of reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reminder'
    post:
      tags:
        - Reminders
      summary: Create a reminder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - petId
                - reminderType
                - title
                - scheduledDate
              properties:
                petId:
                  type: string
                  format: uuid
                reminderType:
                  type: string
                  enum: [vaccine, medication, checkup, appointment, other]
                title:
                  type: string
                  example: "Vacina V10"
                description:
                  type: string
                scheduledDate:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Reminder created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
  /reminders/{id}:
    get:
      tags:
        - Reminders
      summary: Get reminder by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reminder details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
    put:
      tags:
        - Reminders
      summary: Update reminder
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                scheduledDate:
                  type: string
                  format: date-time
                status:
                  type: string
                  enum: [pending, acknowledged, snoozed]
      responses:
        '200':
          description: Reminder updated
    delete:
      tags:
        - Reminders
      summary: Delete reminder
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Reminder deleted
  /reminders/{id}/acknowledge:
    post:
      tags:
        - Reminders
      summary: Acknowledge a reminder
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reminder acknowledged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
  /reminders/{id}/snooze:
    post:
      tags:
        - Reminders
      summary: Snooze a reminder
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                duration:
                  type: integer
                  description: Snooze duration in hours
                  default: 24
      responses:
        '200':
          description: Reminder snoozed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Reminder:
      type: object
      properties:
        id:
          type: string
        petId:
          type: string
          format: uuid
        reminderType:
          type: string
          enum: [vaccine, medication, checkup, appointment, other]
        title:
          type: string
        description:
          type: string
        scheduledDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, sent, acknowledged, snoozed]
        sentAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...